add liveness and readiness check for broker(#1656)

+++ pkg/reconciler/broker/broker_test.go
@@ -857,7 +857,7 @@ func livenessProbe() *corev1.Probe {
 		Handler:             corev1.Handler{
 			HTTPGet: &corev1.HTTPGetAction{
 				Path:        "/healthz",
-				Port:        intstr.IntOrString{Type: intstr.Int, IntVal: int32(8080)},
+				Port:        intstr.IntOrString{Type: intstr.Int, IntVal: 8080},
 			},
 		},
 		InitialDelaySeconds: 5,
@@ -872,7 +872,7 @@ func readinessProbe(containerName string) *corev1.Probe{
 			Handler:             corev1.Handler{
 				HTTPGet: &corev1.HTTPGetAction{
 					Path:        "/readyz",
-					Port:        intstr.IntOrString{Type: intstr.Int, IntVal: int32(8080)},
+					Port:        intstr.IntOrString{Type: intstr.Int, IntVal: 8080},
 				},
 			},
 			InitialDelaySeconds: 5,
