+++ pkg/reconciler/broker/resources/filter.go
@@ -66,7 +66,7 @@ func MakeFilterDeployment(args *FilterArgs) *appsv1.Deployment {
 								Handler:             corev1.Handler{
 									HTTPGet: &corev1.HTTPGetAction{
 										Path:        "/healthz",
-										Port:        intstr.IntOrString{Type: intstr.Int, IntVal: int32(8080)},
+										Port:        intstr.IntOrString{Type: intstr.Int, IntVal: 8080},
 									},
 								},
 								InitialDelaySeconds: 5,
@@ -76,7 +76,7 @@ func MakeFilterDeployment(args *FilterArgs) *appsv1.Deployment {
 								Handler:             corev1.Handler{
 									HTTPGet: &corev1.HTTPGetAction{
 										Path:        "/readyz",
-										Port:        intstr.IntOrString{Type: intstr.Int, IntVal: int32(8080)},
+										Port:        intstr.IntOrString{Type: intstr.Int, IntVal: 8080},
 									},
 								},
 								InitialDelaySeconds: 5,
